Issue 95: username/password for GitHub should be removed - no longer available
With GITHUB no longer needing the username/password, remove these fields from the scaffold but only for GITHUB.

With this change, existing data may now be in an invalid state.

The getGitHubInstance() function has the username/password instance builder removed and now instead throws an exception when the token is not present.
The getRestTemplate() function now only handles "token" and the TokenAuthRestTemplate() already handles throwing exceptions on invalid "token".

It turns out that when this invalid state exists, the service will not start.
This addresses the issue of an unstartable service by preventing the registering of RPMs on start that have the following invalid conditions:
1) If "token" is not available and the type is GITHUB.
2) The "url" is not available and the type is VERSION_ONE.