Re-organize asciidoc index, fix problems, and remove duplicates.
Relocate the cached controller API to the top as they are what is most likely desired.

Remove remoteProjectAPI.adoc and RemoteProjectControllerIntegrationTest.java..

Make sure the index.adoc has all of the appropriate files.

Fix incorrect path, `/status` clearly should instead be `/remote-project-manager`.

Remove the FIXME, this appears to now be working with the removal of RemoteProjectControllerIntegrationTest.java and/or remoteProjectAPI.adoc.

Filenames need to use `API` and not `Api`.